--- os-appium-tests/allure-results/369d7b5a-a810-4ea8-837b-34211aaf93df-attachment.json
@@ -1,4 +1,0 @@
-[
-  "NATIVE_APP",
-  "WEBVIEW_com.outsystems.android"
-]
