--- os-appium-tests/allure-results/687881be-1b11-48e1-bcdd-bd9e6c27d7f1-attachment.json
@@ -1,4 +1,0 @@
-[
-  "NATIVE_APP",
-  "WEBVIEW_com.outsystemsenterprise.ContactsSampleApp"
-]
