--- os-appium-tests/allure-results/96515f94-b98c-4354-8fd3-c186dd80853f-attachment.json
@@ -1,3 +1,0 @@
-{
-  "name": "NATIVE_APP"
-}
