--- os-appium-tests/allure-results/a25f4458-16f9-4411-a992-f45e1206f23f-attachment.json
@@ -1,3 +1,0 @@
-[
-  "NATIVE_APP"
-]
