--- os-appium-tests/allure-results/cb08ad42-fae8-4356-aa67-226e479f472d-attachment.json
@@ -1,3 +1,0 @@
-{
-  "name": "NATIVE_APP"
-}
