--- os-appium-tests/allure-results/cd0dbc7a-b24a-4b54-a249-4d740db1e8a7-testsuite.xml
@@ -1,65 +1,0 @@
-<?xml version='1.0'?>
-<ns2:test-suite xmlns:ns2='urn:model.allure.qatools.yandex.ru' start='1562687436695' stop='1562687440358'>
-    <name>[TestSuite, Description("Add Contact")]</name>
-    <title>[TestSuite, Description("Add Contact")]</title>
-    <test-cases>
-        <test-case start='1562687436698' status='failed' stop='1562687440357'>
-            <name>[Test, Description("Add contact with all parameters"), Priority="P0"]</name>
-            <title>[Test, Description("Add contact with all parameters"), Priority="P0"]</title>
-            <labels>
-                <label name='language' value='javascript'/>
-                <label name='framework' value='wdio'/>
-                <label name='thread' value='0-0'/>
-            </labels>
-            <parameters>
-                <parameter kind='argument' name='device' value='Android'/>
-            </parameters>
-            <steps>
-                <step start='1562687436704' status='passed' stop='1562687436990'>
-                    <name>GET /session/:sessionId/contexts</name>
-                    <title>GET /session/:sessionId/contexts</title>
-                    <attachments>
-                        <attachment title='Response' source='a25f4458-16f9-4411-a992-f45e1206f23f-attachment.json' type='application/json' size='18'/>
-                    </attachments>
-                    <steps/>
-                </step>
-                <step start='1562687437027' status='passed' stop='1562687437486'>
-                    <name>GET /session/:sessionId/contexts</name>
-                    <title>GET /session/:sessionId/contexts</title>
-                    <attachments>
-                        <attachment title='Response' source='369d7b5a-a810-4ea8-837b-34211aaf93df-attachment.json' type='application/json' size='54'/>
-                    </attachments>
-                    <steps/>
-                </step>
-                <step start='1562687437486' status='passed' stop='1562687438229'>
-                    <name>POST /session/:sessionId/context</name>
-                    <title>POST /session/:sessionId/context</title>
-                    <attachments>
-                        <attachment title='Request' source='cb08ad42-fae8-4356-aa67-226e479f472d-attachment.json' type='application/json' size='26'/>
-                    </attachments>
-                    <steps/>
-                </step>
-                <step start='1562687438232' status='passed' stop='1562687440354'>
-                    <name>GET /session/:sessionId/screenshot</name>
-                    <title>GET /session/:sessionId/screenshot</title>
-                    <attachments>
-                        <attachment title='Screenshot' source='8ca99eac-2036-4539-805e-591f0ff8d308-attachment.png' type='image/png' size='225470'/>
-                    </attachments>
-                    <steps/>
-                </step>
-            </steps>
-            <attachments/>
-            <failure>
-                <message>Error: Malformed type for "name" parameter of command switchContext
-                Expected: string
-                Actual: undefined
-                
-                For more info see https://github.com/SeleniumHQ/mobile-spec/blob/master/spec-draft.md#webviews-and-other-contexts
-                </message>
-                <stack-trace>    at &lt;Jasmine>
-                    at Object.switchToContext (C:\Users\iso\Documents\GitHub\cordova-plugin-contacts\os-appium-tests\tests\helpers\Context.ts:41:13)
-                    at UserContext.beforeEach (C:\Users\iso\Documents\GitHub\cordova-plugin-contacts\os-appium-tests\tests\specs\contact.spec.ts:11:17)</stack-trace>
-            </failure>
-        </test-case>
-    </test-cases>
-</ns2:test-suite>
