--- os-appium-tests/allure-results/f2caf532-9184-4834-a61f-caaafb6ff45c-attachment.json
@@ -1,3 +1,0 @@
-{
-  "name": "WEBVIEW_com.outsystemsenterprise.ContactsSampleApp"
-}
