--- os-appium-tests/allure-results/f3d46542-d9d7-4f2b-9d2e-57c7c844437c-attachment.json
@@ -1,3 +1,0 @@
-[
-  "NATIVE_APP"
-]
