+++ os-appium-tests/config/wdio.android.app.conf.js
@@ -8,12 +8,16 @@ const { config } = require('./wdio.shared.conf');
 config.capabilities = [
     {
         // The defaults you need to have in your config
+        automationName: 'UiAutomator2',
         platformName: 'Android',
-        deviceName: 'Android',
+        deviceName: 'Nexus',
         app: join(process.cwd(), 'apps/Contacts_Sample_App.apk'), // Path to your native app
-        // appPackage: 'com.outsystems.android', // Path to your app package
-        // appActivity: 'com.outsystems.android.SplashScreenActivity', // Path to your activity
+        // appPackage: 'com.outsystemsenterprise.enmobiletst.ContactsSampleApp', // Path to your app package
+        // appActivity: 'com.outsystemsenterprise.enmobiletst.ContactsSampleApp.MainActivity', // Path to your activity
+        // chromedriver windows:
         chromedriver: join(process.cwd(), 'chromedriver/chromedriver.exe'),
+        // chromedriver iOS:
+        // chromedriver: join(process.cwd(), 'chromedriver/chromedriver_iOS'),
         noReset: true
     },
 ];
