+++ os-appium-tests/package.json
@@ -21,21 +21,22 @@
   "dependencies": {
     "@types/jasmine": "~3.3.8",
     "@types/node": "~10.12.0",
+    "@wdio/allure-reporter": "~5.7.2",
     "@wdio/cli": "~5.4.19",
     "@wdio/jasmine-framework": "~5.4.18",
     "@wdio/local-runner": "~5.4.19",
     "@wdio/sauce-service": "~5.4.18",
     "@wdio/spec-reporter": "~5.4.15",
     "@wdio/sync": "~5.4.19",
+    "allure-commandline": "~2.9.0",
     "appium": "~1.11.0",
+    "fibers": "^4.0.1",
     "jasmine": "~3.3.1",
     "log4js": "~4.0.2",
     "ts-node": "~7.0.1",
     "typescript": "~3.2.1",
     "wdio-jasmine-framework": "~0.3.8",
     "wdio-spec-reporter": "~0.1.5",
-    "@wdio/allure-reporter": "~5.7.2",
-    "allure-commandline": "~2.9.0",
     "webdriverio": "~5.4.19"
   },
   "devDependencies": {
