+++ os-appium-tests/tests/constants.ts
@@ -1,2 +1,2 @@
-const DEFAULT_TIMEOUT = 10000;
-const DEFAULT_TIMEOUT_INTERVAL = 1000;
+export const DEFAULT_TIMEOUT = 10000;
+export const DEFAULT_TIMEOUT_INTERVAL = 1000;
