+++ os-appium-tests/tests/helpers/Context.ts
@@ -1,4 +1,4 @@
-import "../constants"
+import {DEFAULT_TIMEOUT, DEFAULT_TIMEOUT_INTERVAL} from "../constants"
 
 export const CONTEXT_REF = {
     NATIVE: 'native',
